refactored failAction into a helper method
